Move get_keep to common file.
