+++ cpp/benchmarks/stream_compaction/distinct.cpp
@@ -15,6 +15,7 @@
  */
 
 #include <benchmarks/common/generate_input.hpp>
+#include <benchmarks/stream_compaction/stream_compaction_common.hpp>
 
 #include <cudf/column/column_view.hpp>
 #include <cudf/lists/list_view.hpp>
@@ -26,21 +27,6 @@
 #include <limits>
 
 NVBENCH_DECLARE_TYPE_STRINGS(cudf::timestamp_ms, "cudf::timestamp_ms", "cudf::timestamp_ms");
-
-cudf::duplicate_keep_option get_keep(std::string const& keep_str)
-{
-  if (keep_str == "any") {
-    return cudf::duplicate_keep_option::KEEP_ANY;
-  } else if (keep_str == "first") {
-    return cudf::duplicate_keep_option::KEEP_FIRST;
-  } else if (keep_str == "last") {
-    return cudf::duplicate_keep_option::KEEP_LAST;
-  } else if (keep_str == "none") {
-    return cudf::duplicate_keep_option::KEEP_NONE;
-  } else {
-    CUDF_FAIL("Unsupported keep option.");
-  }
-}
 
 template <typename Type>
 void nvbench_distinct(nvbench::state& state, nvbench::type_list<Type>)
